3. Digit Recognizer/requirements.txt
tensorflow
numpy
matplotlib
scikit-image
streamlit
Pillow
opencv-python